Use long instead of int for coo and related types

+++ 2D/DDLA.cpp
@@ -34,7 +34,7 @@ public:
 	}
 
 	bool f;
-	int n,cursum;
+	long n,cursum;
 	double p;
 	PointQueue pq;
 };
